feat: add PendingPackageRequests component and related styles
