+++ src/features/amUI/userRightsPage/AccessPackageSection/AccessPackageSection.tsx
@@ -13,7 +13,7 @@ import { TabContentSkeleton } from '../../common/RightsTabs/TabContentSkeleton';
 
 import { ActiveDelegations } from './ActiveDelegations';
 import { AccessPackageInfoAlert } from './AccessPackageInfoAlert';
-import { PendingPackageRequests } from './PendingPackageRequests';
+import { PendingPackageRequests } from './PendingPackageRequests/Requests';
 import { QuestionmarkCircleIcon } from '@navikt/aksel-icons';
 
 import classes from './AccessPackageSection.module.css';
