+++ src/features/amUI/userRightsPage/AccessPackageSection/PendingPackageRequests/Requests.module.css
